requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.64.1
grpcio-tools==1.62.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==1.26.4
openai==1.31.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
portalocker==2.8.2
protobuf==4.25.3
pyarrow==16.1.0
pydantic==2.7.3
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.7.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
qdrant-client==1.9.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
setuptools==70.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
streamlit==1.35.0
tenacity==8.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
